SPARK-22255 FileAppender InputStream Read timeout and blocking state
Removed a blocking state.

If there is nothing else on the InputStream and we do a InputStream.read(), we will continue waiting until either new data comes in or there is some sort of timeout. This commit prevents reading an empty InputStream. If the appendToFile conditional is not updated with a check for the number of available. You will have an infinite loop situation appending the same buffer repeatedly to some file because 'n' wont be updated AND we wont be blocking in between each InputStream read and appendToFile.